Move BlockStoreLocation to server module
