+++ servers/src/main/java/tachyon/worker/block/evictor/Evictor.java
@@ -17,7 +17,7 @@ package tachyon.worker.block.evictor;
 
 import java.io.IOException;
 
-import tachyon.worker.BlockStoreLocation;
+import tachyon.worker.block.BlockStoreLocation;
 
 /**
  * Interface for the eviction policy in Tachyon
