+++ servers/src/test/java/tachyon/worker/block/BlockMetadataManagerTest.java
@@ -30,7 +30,6 @@ import com.google.common.collect.Sets;
 
 import tachyon.Constants;
 import tachyon.conf.TachyonConf;
-import tachyon.worker.BlockStoreLocation;
 import tachyon.worker.block.meta.BlockMeta;
 import tachyon.worker.block.meta.StorageDir;
 import tachyon.worker.block.meta.StorageTier;
